refactor: move mocket.compat to mocket.core.compat
