+++ mocket/core/compat.py
