+++ mocket/entry.py
@@ -1,6 +1,6 @@
 import collections.abc
 
-from mocket.compat import encode_to_bytes
+from mocket.core.compat import encode_to_bytes
 from mocket.mocket import Mocket
