+++ mocket/mockhttp.py
@@ -7,7 +7,12 @@ from urllib.parse import parse_qs, unquote, urlsplit
 from h11 import SERVER, Connection, Data
 from h11 import Request as H11Request
 
-from mocket.compat import ENCODING, decode_from_bytes, do_the_magic, encode_to_bytes
+from mocket.core.compat import (
+    ENCODING,
+    decode_from_bytes,
+    do_the_magic,
+    encode_to_bytes,
+)
 from mocket.entry import MocketEntry
 from mocket.mocket import Mocket
