+++ mocket/plugins/httpretty/__init__.py
@@ -1,6 +1,6 @@
 from mocket import mocketize
 from mocket.async_mocket import async_mocketize
-from mocket.compat import ENCODING
+from mocket.core.compat import ENCODING
 from mocket.mocket import Mocket
 from mocket.mockhttp import Entry as MocketHttpEntry
 from mocket.mockhttp import Request as MocketHttpRequest
