+++ mocket/ssl/socket.py
@@ -5,7 +5,7 @@ from datetime import datetime, timedelta
 from ssl import Options
 from typing import Any
 
-from mocket.compat import encode_to_bytes
+from mocket.core.compat import encode_to_bytes
 from mocket.mocket import Mocket
 from mocket.socket import MocketSocket
 from mocket.types import _PeerCertRetDictType
