+++ tests/test_compat.py
@@ -1,4 +1,4 @@
-from mocket.compat import do_the_magic
+from mocket.core.compat import do_the_magic
 
 
 def test_unknown_binary():
